chore: only exposed externally used methods
our dhis-api services interfaces contained some methods that are only
used internally by the service implementation. Lets only expose whats
needed. This makes refactoring easier :sweat_smile:
