+++ dhis-2/dhis-services/dhis-service-core/src/test/java/org/hisp/dhis/trackedentity/DefaultTrackedEntityServiceTest.java
@@ -51,7 +51,10 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.mockito.junit.jupiter.MockitoSettings;
+import org.mockito.quality.Strictness;
 
+@MockitoSettings(strictness = Strictness.LENIENT)
 @ExtendWith(MockitoExtension.class)
 class DefaultTrackedEntityServiceTest {
 
@@ -105,6 +108,8 @@ class DefaultTrackedEntityServiceTest {
     params.setProgram(new Program("Test program"));
     params.getProgram().setMaxTeiCountToReturn(10);
     params.setTrackedEntityUids(Set.of("1"));
+
+    when(aclService.canDataRead(params.getUser(), params.getProgram())).thenReturn(true);
   }
 
   @Test
@@ -119,7 +124,7 @@ class DefaultTrackedEntityServiceTest {
     IllegalQueryException expectedException =
         assertThrows(
             IllegalQueryException.class,
-            () -> teiService.validateSearchScope(params),
+            () -> teiService.getTrackedEntities(params, true, false),
             "test message");
 
     assertEquals("maxteicountreached", expectedException.getMessage());
@@ -134,6 +139,17 @@ class DefaultTrackedEntityServiceTest {
     String currentUsername = CurrentUserUtil.getCurrentUsername();
     when(userService.getUserByUsername(currentUsername)).thenReturn(user);
 
-    teiService.validateSearchScope(params);
+    teiService.getTrackedEntities(params, true, false);
+  }
+
+  @Test
+  void testTeiQueryParamsWithoutEitherProgramOrTrackedEntityType() {
+    TrackedEntityQueryParams params = new TrackedEntityQueryParams();
+
+    IllegalQueryException exception =
+        assertThrows(
+            IllegalQueryException.class, () -> teiService.getTrackedEntities(params, false, true));
+    assertEquals(
+        "Either Program or Tracked entity type should be specified", exception.getMessage());
   }
 }
